saltc: Correct first argument name one of Params constructors in params.h saltc: Changed int arguments in Params constructors to uint saltc: Add explicit conversion to calling Params constructor in main function

+++ saltc/saltc.cpp
@@ -14,7 +14,7 @@ using namespace salt;
  * precompilation, parser, tokenizer, validator and synthesizer.
  */
 int main(int argc, char* argv[]) {
-    arg_parser::Params parameters(argc, argv);
+    arg_parser::Params parameters((uint) argc, argv);
 
     return 0;
 }
